requirements.txt
Pillow>=10.0
python-barcode>=0.15